+++ app/helpers/videos_helper.rb
@@ -1,0 +1,2 @@
+module VideosHelper
+end
