--- app/views/dishes/index.html.erb
@@ -1,22 +1,0 @@
-<div class="my-8 p-4 bg-gray-50 rounded-lg text-center">
-  <h1 class="text-3xl sm:text-4xl font-bold mb-6">動画を探しにいく</h1>
-  <%= link_to '動画を探す', new_dish_path, class: "px-8 py-3 text-white font-semibold bg-indigo-500 rounded-md transition duration-300 hover:bg-indigo-600 focus:outline-none focus:ring focus:ring-indigo-300 focus:ring-opacity-50" %>
-</div>
-
-<div class="flex justify-between items-center">
-  <h1 class="text-2xl font-bold mb-4">料理名の数: <%= @dishes.count %></h1>
-  <h1 class="text-2xl font-bold mb-4">動画の数: <%= @videos.count %></h1>
-</div>
-
-<% @videos.each do |video| %>
-  <div class="flex flex-col md:flex-row md:space-x-4 md:items-center p-4 border-b-2">
-    <div class="md:w-1/3 text-center md:text-left">
-      <h2 class="text-gray-800 bg-gray-200 py-2 px-4 rounded-lg inline-block">
-        <%= video.title %>
-      </h2>
-      <div class="md:w-2/3 text-center">
-        <iframe class="mt-4 mb-4 pb-10 mx-auto md:ml-4 md:mb-0 inline-block" width="280" height="157" src="https://www.youtube.com/embed/<%= video.video_id %>" frameborder=1 allow="fullscreen"></iframe>
-      </div>
-    </div>
-  </div>
-<% end %>
